requirements.txt
pandas
openpyxl
requests
python-dotenv
xlsxwriter
plotly
streamlit